% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{.DollarNames.MyCstruct_extptr}
\alias{.DollarNames.MyCstruct_extptr}
\title{Autocomplete helper for \code{MyCStruct} objects}
\usage{
\method{.DollarNames}{MyCstruct_extptr}(x, pattern)
}
\arguments{
\item{x}{object}

\item{pattern}{current pattern to match}
}
\description{
Autocomplete helper for \code{MyCStruct} objects
}
