% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library-wrapper.R
\name{multiply}
\alias{multiply}
\title{Call the C library function with an externalptr}
\usage{
multiply(mycstruct)
}
\arguments{
\item{mycstruct}{externalptr to a MyCStruct object}
}
\description{
Call the C library function with an externalptr
}
