% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{MyCStruct}
\alias{MyCStruct}
\alias{MyCStructs}
\alias{MyCStruct_copy}
\alias{MyCStruct_set}
\alias{MyCStruct_get}
\alias{$.MyCstruct_extptr}
\alias{[[.MyCstruct_extptr}
\alias{$<-.MyCstruct_extptr}
\alias{[[<-.MyCstruct_extptr}
\alias{print.MyCstruct_extptr}
\alias{as.character.MyCstruct_extptr}
\title{Tools for MyCStruct objects}
\usage{
MyCStruct(N = 0, strength = 0)

MyCStructs(n)

MyCStruct_copy(mycstruct)

MyCStruct_set(mycstruct, field, value)

MyCStruct_get(mycstruct, field)

\method{$}{MyCstruct_extptr}(mycstruct, value)

\method{[[}{MyCstruct_extptr}(mycstruct, value)

\method{$}{MyCstruct_extptr}(mycstruct, field) <- value

\method{[[}{MyCstruct_extptr}(mycstruct, field) <- value

\method{print}{MyCstruct_extptr}(x, ...)

\method{as.character}{MyCstruct_extptr}(x, ...)
}
\arguments{
\item{N}{[unsigned int] Count of objects}

\item{strength}{[double] Some continuous parameters}

\item{n}{number of blank objects to create}

\item{mycstruct}{MyCStruct object}

\item{field}{name of struct field}

\item{value}{value to assign to field}

\item{x, ...}{object and options passed to \code{as.character()}}
}
\description{
Tools for MyCStruct objects
}
\details{
These R functions have been generated from the struct defintion.
}
